requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask_Cors==4.0.0
openai==1.52.2
python-dotenv==1.0.1
vertexai==1.70.0
gunicorn==21.2.0
